pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "activediff"
version = "0.1.0"
description = "Active learning pipeline for nanophotonic design using diffusion models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Vincent", email = "your.email@example.com"}
]
keywords = ["deep-learning", "diffusion-models", "active-learning", "nanophotonics", "pytorch-lightning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "wandb>=0.15.0",
    "timm>=0.9.0",
    "einops>=0.7.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-datadir>=1.4.0",
    "pytest-regressions>=2.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

meep = [
    "meep>=1.25.0",
]

all = [
    "activediff[dev,meep]",
]

[project.urls]
Homepage = "https://github.com/yourusername/activediff"
Repository = "https://github.com/yourusername/activediff"
Documentation = "https://github.com/yourusername/activediff#readme"
Issues = "https://github.com/yourusername/activediff/issues"

[project.scripts]
activediff-train = "activediff.train:main"
activediff-eval = "activediff.eval:main"

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = ["activediff", "activediff.*"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.pytest_cache
  | __pycache__
  | outputs
  | wandb
  | active_learning_output
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "hydra.*",
    "omegaconf.*",
    "pytorch_lightning.*",
    "timm.*",
    "einops.*",
    "PIL.*",
]
ignore_missing_imports = true
